// src/_comp_/lib/types.d.ts
export type FontDecoration = "underline" | "line-through" | "none";

export type FontStyle = "normal" | "italic" | "oblique";

export type FontWeight = "normal" | "bold" | "bolder" | "lighter";

export type HorizontalAlignment = "left" | "right" | "center";

export type IconSymbol3DLayerAnchorOption =
  | "center"
  | "left"
  | "right"
  | "top"
  | "bottom"
  | "top-left"
  | "top-right"
  | "bottom-left"
  | "bottom-right"
  | "relative";

export type IconSymbol3DLayerResourcePrimitiveOption =
  | "circle"
  | "square"
  | "cross"
  | "x"
  | "kite"
  | "triangle";

export type ObjectSymbol3DLayerAnchorOption =
  | "center"
  | "top"
  | "bottom"
  | "origin"
  | "relative";

export type ObjectSymbol3DLayerResourcePrimitiveOption =
  | "sphere"
  | "cylinder"
  | "cube"
  | "cone"
  | "inverted-cone"
  | "diamond"
  | "tetrahedron";

export type VerticalAlignment = "baseline" | "top" | "middle" | "bottom";

export type SimpleLineSymbolCapOption = "butt" | "round" | "square";

export type SimpleLineSymbolJoinOption = "miter" | "round" | "bevel";

export type SimpleLineSymbolStyleOption =
  | "dash"
  | "dash-dot"
  | "dot"
  | "long-dash"
  | "long-dash-dot"
  | "long-dash-dot-dot"
  | "none"
  | "short-dash"
  | "short-dash-dot"
  | "short-dash-dot-dot"
  | "short-dot"
  | "solid";

export type SimpleFillSymbolStyleOption =
  | "backward-diagonal"
  | "cross"
  | "diagonal-cross"
  | "forward-diagonal"
  | "horizontal"
  | "none"
  | "solid"
  | "vertical";

export type SimpleMarkerSymbolStyleOption =
  | "circle"
  | "square"
  | "cross"
  | "x"
  | "diamond"
  | "triangle"
  | "path";

export type LineSymbolMarkerPlacementOption = "begin" | "end" | "begin-end";

export type LineSymbolMarkerStyleOption =
  | "arrow"
  | "circle"
  | "square"
  | "diamond"
  | "cross"
  | "x";